+++ src/components/Widgets/ReviewStatusMetricsWidget/Messages.js
@@ -1,0 +1,15 @@
+import { defineMessages } from 'react-intl'
+
+/**
+ * Internationalized messages for use with TaskCompletionWidget
+ */
+export default defineMessages({
+  label: {
+    id: "Widgets.ReviewStatusMetricsWidget.label",
+    defaultMessage: "Review Status Metrics",
+  },
+  title: {
+    id: "Widgets.ReviewStatusMetricsWidget.title",
+    defaultMessage: "Review Status",
+  }
+})