+++ src/pages/Review/Messages.js
@@ -1,0 +1,21 @@
+import { defineMessages } from 'react-intl'
+
+/**
+ * Internationalized messages for use with TaskAnalysisTable
+ */
+export default defineMessages({
+  tasksToBeReviewed: {
+    id: "Review.Dashboard.tasksToBeReviewed",
+    defaultMessage: "Tasks to be Reviewed",
+  },
+
+  tasksReviewedByMe: {
+    id: "Review.Dashboard.tasksReviewedByMe",
+    defaultMessage: "Tasks Reviewed by Me",
+  },
+
+  myReviewTasks: {
+    id: "Review.Dashboard.myReviewTasks",
+    defaultMessage: "My Reviewed Tasks",
+  },
+})