+++ src/services/Task/TaskReview/TaskReviewNeeded.js
@@ -1,13 +1,12 @@
 import _get from 'lodash/get'
 import _values from 'lodash/values'
-import _uniqueId from 'lodash/uniqueId'
 import _sortBy from 'lodash/sortBy'
 import _reverse from 'lodash/reverse'
 import _snakeCase from 'lodash/snakeCase'
-import format from 'date-fns/format'
 import { defaultRoutes as api } from '../../Server/Server'
 import Endpoint from '../../Server/Endpoint'
 import RequestStatus from '../../Server/RequestStatus'
+import { setupFilterSearchParameters } from './TaskReview'
 import { taskSchema } from '.././Task'
 import { addError } from '../../Error/Error'
 import AppErrors from '../../Error/AppErrors'
@@ -22,12 +21,11 @@ export const RECEIVE_REVIEW_NEEDED_TASKS = 'RECEIVE_REVIEW_NEEDED_TASKS'
  */
 export const receiveReviewNeededTasks = function(tasks,
                                                  status=RequestStatus.success,
-                                                 fetchId, totalCount) {
+                                                 totalCount) {
   return {
     type: RECEIVE_REVIEW_NEEDED_TASKS,
     status,
     tasks,
-    fetchId,
     totalCount,
     receivedAt: Date.now(),
   }
@@ -43,32 +41,11 @@ export const fetchReviewNeededTasks = function(criteria, limit=50) {
   const order = (_get(criteria, 'sortCriteria.direction') || 'DESC').toUpperCase()
   const sort = sortBy ? _snakeCase(sortBy) : null
   const page = _get(criteria, 'page', 0)
-  const filters = _get(criteria, 'filters', {})
-
-  const searchParameters = {}
-  if (filters.reviewRequestedBy) {
-    searchParameters.o = filters.reviewRequestedBy
-  }
-  if (filters.reviewedBy) {
-    searchParameters.r = filters.reviewedBy
-  }
-  if (filters.challenge) {
-    searchParameters.cs = filters.challenge
-  }
-  if (filters.status && filters.status !== "all") {
-    searchParameters.tStatus = filters.status
-  }
-  if (filters.reviewStatus && filters.reviewStatus !== "all") {
-    searchParameters.trStatus = filters.reviewStatus
-  }
-  if (filters.reviewedAt) {
-    searchParameters.startDate = format(filters.reviewedAt, 'YYYY-MM-DD')
-    searchParameters.endDate = format(filters.reviewedAt, 'YYYY-MM-DD')
-  }
+  const searchParameters = setupFilterSearchParameters(_get(criteria, 'filters', {}),
+                                                       criteria.boundingBox,
+                                                       _get(criteria, 'savedChallengesOnly'))
 
   return function(dispatch) {
-    const fetchId = _uniqueId()
-    dispatch(receiveReviewNeededTasks(null, RequestStatus.inProgress, fetchId))
     return new Endpoint(
       api.tasks.review,
       {
@@ -84,11 +61,11 @@ export const fetchReviewNeededTasks = function(criteria, limit=50) {
           tasks = _reverse(tasks)
         }
       }
-      dispatch(receiveReviewNeededTasks(tasks, RequestStatus.success, fetchId,
+      dispatch(receiveReviewNeededTasks(tasks, RequestStatus.success,
                                         normalizedResults.result.total))
       return tasks
     }).catch((error) => {
-      dispatch(receiveReviewNeededTasks([], RequestStatus.error, fetchId))
+      dispatch(receiveReviewNeededTasks([], RequestStatus.error))
       dispatch(addError(AppErrors.reviewTask.fetchFailure))
       console.log(error.response || error)
     })