+++ src/services/Task/TaskReview/TaskReviewed.js
@@ -6,17 +6,15 @@ import { addError } from '../../Error/Error'
 import AppErrors from '../../Error/AppErrors'
 import _get from 'lodash/get'
 import _values from 'lodash/values'
-import _uniqueId from 'lodash/uniqueId'
 import _sortBy from 'lodash/sortBy'
 import _reverse from 'lodash/reverse'
 import _snakeCase from 'lodash/snakeCase'
-import format from 'date-fns/format'
+import { setupFilterSearchParameters } from './TaskReview'
 
 // redux actions
 export const RECEIVE_REVIEWED_TASKS = 'RECEIVE_REVIEWED_TASKS'
 export const RECEIVE_REVIEWED_BY_USER_TASKS = 'RECEIVE_REVIEWED_BY_USER_TASKS'
 
-// redux action creators
 
 /**
  * Add or replace the reviewed tasks in the redux store
@@ -24,13 +22,11 @@ export const RECEIVE_REVIEWED_BY_USER_TASKS = 'RECEIVE_REVIEWED_BY_USER_TASKS'
 export const receiveReviewedTasks = function(tasks,
                                             type,
                                             status=RequestStatus.success,
-                                            fetchId,
                                             totalCount) {
   return {
     type: type,
     status,
     tasks,
-    fetchId,
     totalCount,
     receivedAt: Date.now(),
   }
@@ -47,34 +43,13 @@ export const fetchReviewedTasks = function(criteria, asReviewer, limit=50) {
   const order = (_get(criteria, 'sortCriteria.direction') || 'DESC').toUpperCase()
   const sort = sortBy ? _snakeCase(sortBy) : null
   const page = _get(criteria, 'page', 0)
-  const filters = _get(criteria, 'filters', {})
 
-  const searchParameters = {}
-  if (filters.reviewRequestedBy) {
-    searchParameters.o = filters.reviewRequestedBy
-  }
-  if (filters.reviewedBy) {
-    searchParameters.r = filters.reviewedBy
-  }
-  if (filters.challenge) {
-    searchParameters.cs = filters.challenge
-  }
-  if (filters.status && filters.status !== "all") {
-    searchParameters.tStatus = filters.status
-  }
-  if (filters.reviewStatus && filters.reviewStatus !== "all") {
-    searchParameters.trStatus = filters.reviewStatus
-  }
-  if (filters.reviewedAt) {
-    searchParameters.startDate = format(filters.reviewedAt, 'YYYY-MM-DD')
-    searchParameters.endDate = format(filters.reviewedAt, 'YYYY-MM-DD')
-  }
+  const searchParameters = setupFilterSearchParameters(_get(criteria, 'filters', {}), criteria.boundingBox)
 
   return function(dispatch) {
-    const fetchId = _uniqueId()
     dispatch(receiveReviewedTasks(null,
       asReviewer ? RECEIVE_REVIEWED_BY_USER_TASKS: RECEIVE_REVIEWED_TASKS,
-      RequestStatus.inProgress, fetchId))
+      RequestStatus.inProgress))
     return new Endpoint(
       api.tasks.reviewed,
       {
@@ -93,12 +68,12 @@ export const fetchReviewedTasks = function(criteria, asReviewer, limit=50) {
 
       dispatch(receiveReviewedTasks(tasks,
         asReviewer ? RECEIVE_REVIEWED_BY_USER_TASKS : RECEIVE_REVIEWED_TASKS,
-        RequestStatus.success, fetchId, normalizedResults.result.total))
+        RequestStatus.success, normalizedResults.result.total))
       return tasks
     }).catch((error) => {
       dispatch(receiveReviewedTasks([],
         asReviewer ? RECEIVE_REVIEWED_BY_USER_TASKS: RECEIVE_REVIEWED_TASKS,
-        RequestStatus.error, fetchId))
+        RequestStatus.error))
       dispatch(addError(AppErrors.reviewTask.fetchFailure))
       console.log(error.response || error)
     })